+++ android/src/main/kotlin/com/inappstory/inappstory_plugin/adaptors/IASManagerAdaptor.kt
@@ -36,6 +36,10 @@ class IASManagerAdaptor(
             true, ForceCloseReaderCallback {})
     }
 
+    override fun clearCache() {
+        inAppStoryManager.clearCache()
+    }
+
     override fun setTransparentStatusBar() {
         inAppStoryManager.setOpenStoriesReader(CustomOpenStoriesReader())
     }