+++ android/src/main/kotlin/com/inappstory/inappstory_plugin/adaptors/IASSdkModuleAdaptor.kt
@@ -47,6 +47,7 @@ class InappstorySdkModuleAdaptor(
         apiKey: String, userID: String, callback: (Result<Unit>) -> Unit
     ) {
         try {
+            feedListAdaptors.clear()
             inAppStoryManager = inAppStoryAPI.inAppStoryManager.create(
                 apiKey,
                 userID,