+++ ios/Classes/Adaptors/IASStatisticsManagerAdaptor.swift
@@ -4,15 +4,15 @@ import Foundation
 @_spi(IAS_API) import InAppStorySDK
 
 class IASStatisticsManagerAdaptor: InAppStoryStatManagerHostApi {
-    func sendStatistics(enabled: Bool, completion: @escaping (Result<Void, any Error>) -> Void) {
+    func sendStatistics(
+        enabled: Bool,
+        completion: @escaping (Result<Void, any Error>) -> Void
+    ) {
         InAppStory.shared.isStatisticDisabled = !enabled
-        let userId: String = InAppStory.shared.settings?.userID ?? ""
-        
-        InAppStory.shared.settings = Settings(userID: userId)
 
         completion(.success(()))
     }
-    
+
     init(binaryMessenger: FlutterBinaryMessenger) {
         InAppStoryStatManagerHostApiSetup.setUp(
             binaryMessenger: binaryMessenger,